Move LocalizedConsoleKeys to Sandra.UI.csproj and DisplayShortcutKeys method to CombinedUIActionInterface.
